+++ R/cpos_method.R
@@ -62,7 +62,7 @@ setMethod("cpos", "character", function(.Object, query, pAttribute = getOption("
       }
     )
     hits <- do.call(rbind, hitList)
-  } else if (cqp == TRUE) {
+  } else if (cqp) {
     hitList <- lapply(
       query,
       function(Q){
@@ -82,7 +82,7 @@ setMethod("cpos", "character", function(.Object, query, pAttribute = getOption("
     )
     hits <- do.call(rbind, hitList)
   }
-  hits
+  if (nrow(hits) == 0) invisible(NULL) else hits
 })
   
 #' @rdname cpos-method
@@ -103,7 +103,7 @@ setMethod("cpos", "partition", function(.Object, query, cqp = is.cqp, pAttribute
 #'   the tempcorpus will be shifted so that they match the positions of the
 #'   corpus from which the tempcorpus was generated
 #' @rdname cpos-method
-setMethod("cpos", "tempcorpus", function(.Object, query, shift=TRUE){
+setMethod("cpos", "tempcorpus", function(.Object, query, shift = TRUE){
   cqpBatchCmds <- paste(paste(c(
     "TMPCORPUS",
     paste("FOO = ", query, sep=""),
@@ -117,18 +117,18 @@ setMethod("cpos", "tempcorpus", function(.Object, query, shift=TRUE){
   cpos <- system(cqpCmd, intern=TRUE)
   cposMatrix <- do.call(rbind, lapply(strsplit(cpos, "\t"), as.integer))
   cposMatrix <- cposMatrix[,1:2]
-  if (shift == TRUE) cposMatrix <- apply(cposMatrix, 2, function(x) x + .Object@cpos[1,1])
+  if (shift) cposMatrix <- apply(cposMatrix, 2, function(x) x + .Object@cpos[1,1])
   cposMatrix
 })
 
 #' @rdname cpos-method
-setMethod("cpos", "matrix", function(.Object){
-  as.vector(unlist(apply(.Object, 1, function(row) c(row[1]:row[2]))))  
-})
+setMethod("cpos", "matrix", function(.Object)
+  as.vector(unlist(apply(.Object, 1, function(row) row[1]:row[2])))  
+)
 
 #' @rdname cpos-method
-setMethod("cpos", "hits", function(.Object){
+setMethod("cpos", "hits", function(.Object)
   cpos(as.matrix(.Object@stat[, c("cpos_left", "cpos_right")]))
-})
+)