+++ inst/shiny/modules/corpus.R
@@ -20,7 +20,7 @@ corpusServer <- function(input, output, session){
   observeEvent(
     input$corpus_corpus,
     {
-      infoFile <- RegistryFile$new(input$corpus_corpus)$getInfo()
+      infoFile <- registry_get_info(input$corpus_corpus)
       if (file.exists(infoFile)){
         content <- readLines(infoFile)
         if (grepl(".md$", infoFile)){